+++ src/userpw.c
@@ -37,7 +37,7 @@ int		userpw_entry(int argc, char** argv, int argo, commandEntry_t * entry)
 	bool				hexOutput = false;
 	char *				password = NULL;
 	char				hash[MAX_DIGEST_SIZE];
-	uint32_t			hashLen = sizeof(hash);
+	size_t				hashLen = sizeof(hash);
 	char				hex[(sizeof(hash) * 2) + 1];
 	size_t				hexLen = 0;
 	char *				out;
